ai_service/requirements.txt
# AI Recovery Plan Microservice - Requirements
# Python 3.10 environment

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
tensorflow-cpu==2.12.0  # CPU-only version for smaller footprint
numpy==1.23.5
scikit-learn==1.3.2
joblib==1.3.2

# HTTP Client (for testing)
requests==2.31.0

# Utilities
python-multipart==0.0.6
